Move affine2orientation function to helper_functions and update tests accordingly.
